requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
eventlet==0.40.0
Flask==3.1.0
Flask-SocketIO==5.5.1
fonttools==4.57.0
greenlet==3.2.2
idna==3.10
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
nltk==3.9.1
numpy==2.2.5
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-engineio==4.12.1
python-socketio==5.13.0
regex==2024.11.6
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
tqdm==4.67.1
urllib3==2.4.0
Werkzeug==3.1.3
keybert
flask_sqlalchemy==3.1.1
python-dotenv
pymysql==1.1.0
flask_cors==3.0.10
flask-jwt-extended==4.6.0
flask-migrate==4.0.5
rank_bm25
rapidfuzz
sentence-transformers
scikit-learn
transformers
torch
